+++ CHANGELOG.md
@@ -23,6 +23,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
     - `hasSizeNotEqualTo`
     - `isEmpty` / `isNotEmpty`
     - `isContaining` / `isNotContaining`
+  - `Map`
+    - `isContainingKey` / `isNotContainingKey`
+    - `isContainingValue` / `isNotContainingValue`
 
 ### Changed